refactor: add getLinkArrowLabel to the common package
HH-283
