+++ apps/events-helsinki/src/hooks/useRHHCConfig.tsx
@@ -1,13 +1,10 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import type { EventFields } from 'events-helsinki-components';
 import {
+  getLinkArrowLabel,
   useLocale,
   useCommonTranslation,
   useCmsTranslation,
-  useEventTranslation,
   MAIN_CONTENT_ID,
-  getEnrolmentStatus,
-  getDateRangeStr,
 } from 'events-helsinki-components';
 import Head from 'next/head';
 import Link from 'next/link';
@@ -39,7 +36,6 @@ export default function useRHHCConfig() {
   const apolloClient = useApolloClient();
   const { t: commonTranslation } = useCommonTranslation();
   const { t: cmsTranslation } = useCmsTranslation();
-  const { t: eventTranslation } = useEventTranslation();
   const locale = useLocale();
 
   return React.useMemo(() => {
@@ -126,28 +122,9 @@ export default function useRHHCConfig() {
       },
       utils: {
         ...rhhcDefaultConfig.utils,
-        getEventCardProps: (item: EventFields, locale: string) => {
-          const status = getEnrolmentStatus(item);
-          const linkArrowLabel = eventTranslation(
-            `event:enrolmentStatus.${status}`,
-            {
-              date: item.enrolmentStartTime
-                ? getDateRangeStr({
-                    start: item.enrolmentStartTime,
-                    locale,
-                    includeWeekday: false,
-                    includeTime: true,
-                    timeAbbreviation: commonTranslation(
-                      'common:timeAbbreviation'
-                    ),
-                  })
-                : '',
-            }
-          );
-          return {
-            ...rhhcDefaultConfig.utils.getEventCardProps(item, locale),
-            linkArrowLabel,
-          };
+        getEventCardProps: {
+          ...rhhcDefaultConfig.utils.getEventCardProps,
+          getLinkArrowLabel,
         },
         getRoutedInternalHref: (
           link: string,
@@ -175,5 +152,5 @@ export default function useRHHCConfig() {
       },
       internalHrefOrigins,
     } as unknown as Config;
-  }, [commonTranslation, cmsTranslation, apolloClient, locale]);
+  }, [commonTranslation, locale, apolloClient, cmsTranslation]);
 }
